chore: refine index file
