+++ lib/index.ts
@@ -1,4 +1,6 @@
+export * from './apiClient'
 export * from './constants'
 export * from './dayjsX'
+export * from './storage'
 export * from './theme'
 export * from './utils'
