+++ store/global/index.ts
@@ -1,6 +1,6 @@
+import { zustandStorage } from '@/lib'
 import { create } from 'zustand'
 import { createJSONStorage, persist } from 'zustand/middleware'
-import { zustandStorage } from '../storage'
 import { createLanguageSlice } from './language'
 import { GlobalStore } from './shared.types'
 import { createThemeSlice } from './theme'
